requirements.txt
nextcord
python-dotenv
Flask[async]